+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-22](https://github.com/OS2web/os2web_datalookup/pull/22)
+  Set actual identifier from lookup on lookup result
+
 ## [3.4.1] - 2026-08-14
   Drupal 11 compatibility fixes